#434: Renamed the file containing the SkillModifierIncreaseEffect
 -
